src/src/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
pydantic
python-dotenv
 requests
